Change validation to be package-based
