--- home_assistant_config_validator/config/validation.py
@@ -1,216 +1,0 @@
-"""Configuration for how to validate each Package."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import ClassVar, Literal, TypedDict
-
-from wg_utilities.functions.json import JSONArr, JSONObj, JSONVal
-
-from home_assistant_config_validator.utils import (
-    check_known_entity_usages,
-    const,
-    load_yaml,
-)
-
-from .base import Config, replace_non_alphanumeric
-from .parser import ParserConfig
-
-
-class ShouldMatchFilepathItem(TypedDict):
-    """Type definition for a single item in the `should_match_filepath` list."""
-
-    field: str
-    separator: str
-    prefix: str
-    ignore_chars: str
-    include_domain_dir: bool
-    remove_sensor_prefix: bool
-
-
-@dataclass
-class ValidationConfig(Config):
-    """Dataclass for a domain's validator configuration."""
-
-    CONFIGURATION_TYPE: ClassVar[Literal[const.ConfigurationType.VALIDATION]] = (
-        const.ConfigurationType.VALIDATION
-    )
-
-    package_name: str
-
-    should_match_filename: list[str] = field(default_factory=list)
-    should_match_filepath: list[ShouldMatchFilepathItem] = field(
-        default_factory=list,
-    )
-    should_be_equal: list[tuple[str, ...]] = field(default_factory=list)
-    should_exist: list[str] = field(default_factory=list)
-    should_be_hardcoded: dict[str, JSONVal] = field(default_factory=dict)
-
-    _package_issues: dict[str, list[Exception]] = field(default_factory=dict)
-
-    def run_custom_validations(  # noqa: PLR0912
-        self,
-        *,
-        domain_dir_path: Path,
-        file_path: Path,
-        entity_yaml: JSONObj,
-    ) -> list[Exception]:
-        # pylint: disable=too-many-locals,too-many-branches
-        """Run all validations on the given entity.
-
-        Args:
-            domain_dir_path (Path): The domain's directory path
-            file_path (Path): The entity's file path
-            entity_yaml (JSONObj): The entity's YAML
-
-        Returns:
-            list[Exception]: A list of exceptions raised during validation
-        """
-        custom_issues: list[Exception] = []
-
-        for smfn_field in self.should_match_filename:
-            # Some entity types (e.g. sensor.systemmonitor) don't have certain fields
-            # (e.g. name). If it's required, it'll get picked up in the other checks.
-            if smfn_field not in entity_yaml:
-                continue
-
-            if (
-                fmt_value := replace_non_alphanumeric(
-                    field_value := entity_yaml.get(  # type: ignore[arg-type]
-                        smfn_field,
-                        "",
-                    ),
-                )
-            ) != file_path.with_suffix("").name.lower():
-                custom_issues.append(
-                    ValueError(
-                        f"`{smfn_field}: {field_value!s}` ({fmt_value=}) should match"
-                        " file name",
-                    ),
-                )
-
-        for field_name_1, field_name_2 in self.should_be_equal:
-            if (field_value_1 := entity_yaml.get(field_name_1)) != (
-                field_value_2 := entity_yaml.get(field_name_2)
-            ):
-                custom_issues.append(
-                    ValueError(
-                        f"`{field_name_1}: {field_value_1!s}` should match"
-                        f" `{field_name_2}: {field_value_2!s}`",
-                    ),
-                )
-
-        for smfp_config in self.should_match_filepath:
-            remove_sensor_prefix = smfp_config.get("remove_sensor_prefix", False)
-
-            if smfp_config.get("include_domain_dir", False):
-                if remove_sensor_prefix:
-                    raise ValueError(  # noqa: TRY003
-                        "include_domain_dir and remove_sensor_prefix are mutually"
-                        " exclusive",
-                    )
-
-                filepath_parts = (
-                    file_path.with_suffix("").relative_to(domain_dir_path.parent).parts
-                )
-            else:
-                filepath_parts = (
-                    file_path.with_suffix("").relative_to(domain_dir_path).parts
-                )
-
-            if remove_sensor_prefix and filepath_parts[0] in (
-                "binary_sensor",
-                "sensor",
-            ):
-                filepath_parts = filepath_parts[1:]
-
-            expected_value = smfp_config.get("prefix", "") + (
-                sep := smfp_config["separator"]
-            ).join(filepath_parts)
-
-            actual_value = replace_non_alphanumeric(
-                entity_yaml.get(smfp_config["field"], ""),  # type: ignore[arg-type]
-                ignore_chars=sep.replace(" ", ""),
-            )
-
-            if sep == " ":
-                actual_value = actual_value.replace("_", " ")
-                expected_value = expected_value.replace("_", " ")
-
-            if expected_value != actual_value:
-                custom_issues.append(
-                    ValueError(
-                        f"`{smfp_config['field']}: {(actual_value or 'null')!s}`"
-                        f" should match file path: `{expected_value!s}`",
-                    ),
-                )
-
-        for se_field in self.should_exist:
-            if se_field not in entity_yaml:
-                custom_issues.append(ValueError(f"`{se_field}` should be defined"))
-
-        for sbh_field, hardcoded_value in self.should_be_hardcoded.items():
-            if (field_value := entity_yaml.get(sbh_field)) != hardcoded_value:
-                custom_issues.append(
-                    ValueError(
-                        f"`{sbh_field}: {field_value!s}` should be hardcoded as"
-                        f" {hardcoded_value!r}",
-                    ),
-                )
-
-        return custom_issues
-
-    def validate_domain(self) -> None:
-        """Validate a domain's YAML files.
-
-        Each file is validated against Home Assistant's schema for that domain, and
-        against the domain's validator configuration. Custom validation is done by
-        evaluating the JSON configuration adjacent to this module and applying each
-        validation rule to the entity.
-
-        Invalid files are added to the `self._domain_issues` dict, with the file path
-        as the key and a list of exceptions as the value.
-        """
-        domain_dir_path = const.ENTITIES_DIR / self.package_name
-
-        if not domain_dir_path.is_dir():
-            self._package_issues["root"] = [
-                ValueError(
-                    f"Directory {domain_dir_path.relative_to(const.REPO_PATH)} does not exist",
-                ),
-            ]
-
-        parser = ParserConfig.get_for_package(self.package_name)
-
-        for file in domain_dir_path.rglob("*.yaml"):
-            entity_yaml: JSONObj | JSONArr = parser.parse(
-                load_yaml(file, resolve_tags=False),
-            )
-
-            file_issues: list[Exception] = self.run_custom_validations(
-                domain_dir_path=domain_dir_path,
-                file_path=file,
-                entity_yaml=entity_yaml,  # type: ignore[arg-type]
-            )
-
-            file_issues.extend(
-                check_known_entity_usages(
-                    entity_yaml,
-                    entity_keys=("entity_id", "service"),
-                ),
-            )
-
-            if file_issues:
-                self._package_issues[file.relative_to(const.REPO_PATH).as_posix()] = (
-                    file_issues
-                )
-
-    @property
-    def package_issues(self) -> dict[str, list[Exception]]:
-        """Return a list of all issues found in the domain.
-
-        Returns:
-            dict[str, list[Exception]]: A list of all issues found in the domain
-        """
-        return self._package_issues
