+++ home_assistant_config_validator/models/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from .package import Package
+from .readme_entity import ReadmeEntity
+
+__all__ = [
+    "Package",
+    "ReadmeEntity",
+]
