+++ pyproject.toml
@@ -17,6 +17,7 @@ python = "^3.11"
 wg-utilities = { extras = ["logging"], version = "^5.8.3" }
 pytz = "^2024.1"
 pyyaml = "^6.0"
+jsonpath-ng = "^1.6.1"
 
 [tool.poetry.group.dev.dependencies]
 pylint-strict-informational = "^0.1"
